refactor(common): clean up toml config methods
