+++ src/common/src/config.rs
@@ -22,43 +22,23 @@ use crate::error::{Result, RwError};
 
 pub const MAX_CONNECTION_WINDOW_SIZE: u32 = (1 << 31) - 1;
 
-/// TODO(TaoWu): The configs here may be preferable to be managed under corresponding module
-/// separately.
-#[derive(Clone, Debug, Serialize, Deserialize, Default)]
-#[serde(deny_unknown_fields)]
-pub struct ComputeNodeConfig {
-    // For connection
-    #[serde(default)]
-    pub server: ServerConfig,
-
-    // Below for batch query.
-    #[serde(default)]
-    pub batch: BatchConfig,
-
-    // Below for streaming.
-    #[serde(default)]
-    pub streaming: StreamingConfig,
-
-    // Below for Hummock.
-    #[serde(default)]
-    pub storage: StorageConfig,
-}
-
-pub fn load_config(path: &str) -> ComputeNodeConfig {
+pub fn load_config<S>(path: &str) -> Result<S>
+where
+    for<'a> S: Deserialize<'a> + Default,
+{
     if path.is_empty() {
         tracing::warn!("risingwave.toml not found, using default config.");
-        return ComputeNodeConfig::default();
+        return Ok(S::default());
     }
-
-    ComputeNodeConfig::init(path.to_owned().into()).unwrap()
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, Default)]
-#[serde(deny_unknown_fields)]
-pub struct FrontendConfig {
-    // For connection
-    #[serde(default)]
-    pub server: ServerConfig,
+    let config_str = fs::read_to_string(PathBuf::from(path.to_owned())).map_err(|e| {
+        RwError::from(InternalError(format!(
+            "failed to open config file '{}': {}",
+            path, e
+        )))
+    })?;
+    let config: S = toml::from_str(config_str.as_str())
+        .map_err(|e| RwError::from(InternalError(format!("parse error {}", e))))?;
+    Ok(config)
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
@@ -211,36 +191,6 @@ pub struct FileCacheConfig {
 impl Default for FileCacheConfig {
     fn default() -> Self {
         toml::from_str("").unwrap()
-    }
-}
-
-impl ComputeNodeConfig {
-    pub fn init(path: PathBuf) -> Result<ComputeNodeConfig> {
-        let config_str = fs::read_to_string(path.clone()).map_err(|e| {
-            RwError::from(InternalError(format!(
-                "failed to open config file '{}': {}",
-                path.to_string_lossy(),
-                e
-            )))
-        })?;
-        let config: ComputeNodeConfig = toml::from_str(config_str.as_str())
-            .map_err(|e| RwError::from(InternalError(format!("parse error {}", e))))?;
-        Ok(config)
-    }
-}
-
-impl FrontendConfig {
-    pub fn init(path: PathBuf) -> Result<Self> {
-        let config_str = fs::read_to_string(path.clone()).map_err(|e| {
-            RwError::from(InternalError(format!(
-                "failed to open config file '{}': {}",
-                path.to_string_lossy(),
-                e
-            )))
-        })?;
-        let config: FrontendConfig = toml::from_str(config_str.as_str())
-            .map_err(|e| RwError::from(InternalError(format!("parse error {}", e))))?;
-        Ok(config)
     }
 }
