+++ src/frontend/Cargo.toml
@@ -39,6 +39,8 @@ risingwave_rpc_client = { path = "../rpc_client" }
 risingwave_source = { path = "../source" }
 risingwave_sqlparser = { path = "../sqlparser" }
 risingwave_storage = { path = "../storage" }
+serde = { version = "1", features = ["derive"] }
+serde_derive = "1"
 serde_json = "1"
 sha2 = "0.10.2"
 smallvec = { version = "1.6.1", features = ["serde"] }
