+++ src/frontend/src/lib.rs
@@ -65,6 +65,7 @@ use std::sync::Arc;
 
 use clap::Parser;
 use pgwire::pg_server::pg_serve;
+use serde::{Deserialize, Serialize};
 use session::SessionManagerImpl;
 
 #[derive(Parser, Clone, Debug)]
@@ -98,6 +99,8 @@ impl Default for FrontendOpts {
 use std::future::Future;
 use std::pin::Pin;
 
+use risingwave_common::config::ServerConfig;
+
 /// Start frontend
 pub fn start(opts: FrontendOpts) -> Pin<Box<dyn Future<Output = ()> + Send>> {
     // WARNING: don't change the function signature. Making it `async fn` will cause
@@ -106,4 +109,12 @@ pub fn start(opts: FrontendOpts) -> Pin<Box<dyn Future<Output = ()> + Send>> {
         let session_mgr = Arc::new(SessionManagerImpl::new(&opts).await.unwrap());
         pg_serve(&opts.host, session_mgr).await.unwrap();
     })
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, Default)]
+#[serde(deny_unknown_fields)]
+pub struct FrontendConfig {
+    // For connection
+    #[serde(default)]
+    pub server: ServerConfig,
 }
