+++ src/meta/src/lib.rs
@@ -57,7 +57,7 @@ use std::time::Duration;
 
 use clap::{ArgEnum, Parser};
 pub use error::{MetaError, MetaResult};
-use risingwave_common::config::ComputeNodeConfig;
+use serde::{Deserialize, Serialize};
 
 use crate::manager::MetaOpts;
 use crate::rpc::server::{rpc_serve, AddressInfo, MetaStoreBackend};
@@ -149,20 +149,18 @@ pub struct MetaNodeOpts {
     enable_committed_sst_sanity_check: bool,
 }
 
-fn load_config(opts: &MetaNodeOpts) -> ComputeNodeConfig {
-    risingwave_common::config::load_config(&opts.config_path)
-}
-
 use std::future::Future;
 use std::pin::Pin;
 
-/// Start meta node
+use risingwave_common::config::{load_config, StreamingConfig};
 
+/// Start meta node
 pub fn start(opts: MetaNodeOpts) -> Pin<Box<dyn Future<Output = ()> + Send>> {
     // WARNING: don't change the function signature. Making it `async fn` will cause
     // slow compile in release mode.
     Box::pin(async move {
-        let compute_config = load_config(&opts);
+        let meta_config: MetaNodeConfig = load_config(&opts.config_path).unwrap();
+        tracing::info!("Starting meta node with config {:?}", meta_config);
         let meta_addr = opts.host.unwrap_or_else(|| opts.listen_addr.clone());
         let listen_addr = opts.listen_addr.parse().unwrap();
         let dashboard_addr = opts.dashboard_host.map(|x| x.parse().unwrap());
@@ -183,9 +181,9 @@ pub fn start(opts: MetaNodeOpts) -> Pin<Box<dyn Future<Output = ()> + Send>> {
         };
         let max_heartbeat_interval = Duration::from_millis(opts.max_heartbeat_interval as u64);
         let checkpoint_interval =
-            Duration::from_millis(compute_config.streaming.checkpoint_interval_ms as u64);
+            Duration::from_millis(meta_config.streaming.checkpoint_interval_ms as u64);
         let max_idle_ms = opts.dangerous_max_idle_secs.unwrap_or(0) * 1000;
-        let in_flight_barrier_nums = compute_config.streaming.in_flight_barrier_nums as usize;
+        let in_flight_barrier_nums = meta_config.streaming.in_flight_barrier_nums as usize;
 
         tracing::info!("Meta server listening at {}", listen_addr);
         let add_info = AddressInfo {
@@ -217,4 +215,12 @@ pub fn start(opts: MetaNodeOpts) -> Pin<Box<dyn Future<Output = ()> + Send>> {
         join_handle.await.unwrap();
         tracing::info!("Meta server is stopped");
     })
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, Default)]
+#[serde(deny_unknown_fields)]
+pub struct MetaNodeConfig {
+    // Below for streaming.
+    #[serde(default)]
+    pub streaming: StreamingConfig,
 }
