+++ src/storage/compactor/src/lib.rs
@@ -30,13 +30,8 @@ mod compactor_observer;
 mod rpc;
 mod server;
 
-use std::fs;
-use std::path::PathBuf;
-
 use clap::Parser;
 use risingwave_common::config::{ServerConfig, StorageConfig};
-use risingwave_common::error::ErrorCode::InternalError;
-use risingwave_common::error::{Result, RwError};
 use serde::{Deserialize, Serialize};
 
 use crate::server::compactor_serve;
@@ -89,21 +84,6 @@ pub struct CompactorConfig {
     // Below for Hummock.
     #[serde(default)]
     pub storage: StorageConfig,
-}
-
-impl CompactorConfig {
-    pub fn init(path: PathBuf) -> Result<CompactorConfig> {
-        let config_str = fs::read_to_string(path.clone()).map_err(|e| {
-            RwError::from(InternalError(format!(
-                "failed to open config file '{}': {}",
-                path.to_string_lossy(),
-                e
-            )))
-        })?;
-        let config: CompactorConfig = toml::from_str(config_str.as_str())
-            .map_err(|e| RwError::from(InternalError(format!("parse error {}", e))))?;
-        Ok(config)
-    }
 }
 
 use std::future::Future;
